requirements.txt
# Core dependencies for ArmPi Mini Robot
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0

# YAML handling
PyYAML>=6.0

# Serial communication for robot control
pyserial>=3.5

# Raspberry Pi specific (ARM-based boards)
RPi.GPIO>=0.7.0; sys_platform == "linux" and platform_machine in ("armv7l", "aarch64")
picamera2>=0.3.0; sys_platform == "linux" and platform_machine in ("armv7l", "aarch64")

# Computer vision and AI
mediapipe>=0.10.0

# Web server and RPC for remote control
werkzeug>=2.3.0
jsonrpc2-zeromq>=0.5.0

# Additional utilities
Pillow>=10.0.0
scipy>=1.11.0

# For Jetson Nano compatibility
jetson-stats>=4.0.0; platform_machine == "aarch64" and "jetson" in platform.platform().lower()
